+++ packages/core/src/member-profile.server.ts
@@ -27,7 +27,9 @@ export { getEvent } from './modules/event/queries/get-event';
 export { listEventAttendees } from './modules/event/queries/list-event-attendees';
 export { isFeatureFlagEnabled } from './modules/feature-flag/queries/is-feature-flag-enabled';
 export { initializeFeatureFlagServer } from './modules/feature-flag/use-cases/initialize-feature-flag-server';
+export { getPointsLeaderboard } from './modules/gamification/queries/get-points-leaderboard';
 export { getTotalPoints } from './modules/gamification/queries/get-total-points';
+export { listActivities } from './modules/gamification/queries/list-activities';
 export { getIcebreakerPrompts } from './modules/icebreaker/queries/get-icebreaker-prompts';
 export { getIcebreakerResponses } from './modules/icebreaker/queries/get-icebreaker-responses';
 export { upsertIcebreakerResponses } from './modules/icebreaker/use-cases/upsert-icebreaker-responses';
